added authorizations for edit and delete

+++ controllers/feed.js
@@ -156,6 +156,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Erro('Not Authorized');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -185,6 +190,12 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
       //check loggedIn user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Erro('Not Authorized');
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
